+++ crates/core/Cargo.toml
@@ -26,5 +26,6 @@ workspace = true
 base64ct.workspace = true
 fastrand = "2.1.0"
 serde.workspace = true
+serde_json.workspace = true
 
 [dev-dependencies]
